# crates/api/cheetah-http-api/openapi/signaling-v1.yaml
openapi: 3.1.0
info:
  title: Cheetah Signaling API
  version: 1.0.0
  description: Northbound REST API for Cheetah Signaling.
servers:
  - url: /
paths:
  /health/live:
    get:
      operationId: getLiveness
      responses:
        '200':
          description: Service is alive.
  /health/ready:
    get:
      operationId: getReadiness
      responses:
        '200':
          description: Service is ready.
  /api/v1/tenants:
    get:
      operationId: listTenants
      parameters:
        - name: page_size
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
        - name: name_prefix
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A page of tenants.
    post:
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tenant created.
        '200':
          description: Idempotent hit of an existing tenant.
  /api/v1/nodes:
    get:
      operationId: listNodes
      parameters:
        - name: page_size
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A page of alive signaling nodes.
  /api/v1/media-nodes:
    get:
      operationId: listMediaNodes
      parameters:
        - name: page_size
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A page of alive media nodes.
  /api/v1/devices:
    get:
      operationId: listDevices
      parameters:
        - name: page_size
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A page of devices.
    post:
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Device created.
  /api/v1/devices/{id}:
    get:
      operationId: getDevice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device details.
  /api/v1/operations:
    post:
      operationId: createOperation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '202':
          description: Operation accepted.
  /api/v1/operations/{id}:
    get:
      operationId: getOperation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Operation details.
  /api/v1/media/sessions:
    get:
      operationId: listMediaSessions
      parameters:
        - name: page_size
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A page of media sessions.
    post:
      operationId: createMediaSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '202':
          description: Media session accepted.
  /api/v1/media/sessions/{id}:
    get:
      operationId: getMediaSession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Media session details.
        '404':
          description: Media session not found.
  /api/v1/events/stream:
    get:
      operationId: eventStream
      responses:
        '200':
          description: SSE event stream.
  /api/v1/webhooks:
    get:
      operationId: listWebhooks
      parameters:
        - name: page_size
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: event_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A page of webhook configurations.
    post:
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Webhook configuration created.
  /api/v1/webhooks/{id}:
    get:
      operationId: getWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Webhook configuration details.
    patch:
      operationId: updateWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook configuration updated.
    delete:
      operationId: deleteWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Webhook configuration deleted.
  /api/v1/webhooks/{id}/deliveries:
    get:
      operationId: listDeliveries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page_size
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A page of webhook deliveries.
    post:
      operationId: triggerDelivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Delivery triggered.
  /api/v1/webhooks/{webhook_id}/deliveries/{delivery_id}/replay:
    post:
      operationId: replayDelivery
      parameters:
        - name: webhook_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: delivery_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Delivery replayed.
components:
  schemas: {}